choose-your-own-adventure/gpt_api/requirements.txt
openai==0.27.8
flask==2.3.2
python-dotenv==1.0.0
mysql-connector==2.2.9
requests==2.31.0
peewee==3.16.2
typing-extensions==4.7.1

# For serpapi
google-search-results==2.4.2
